show duckdb extensions

show duckdb extensions

Usage

duckdb_extensions(conn = cached_connection())

Arguments

conn

A connection to a database.

Value

a data frame listing all available extensions, with boolean columns indicating which extensions are installed or loaded, and a description of each extension.

Examples

duckdb_extensions()
#>      extension_name loaded installed
#> 1      autocomplete  FALSE     FALSE
#> 2               aws  FALSE     FALSE
#> 3             azure  FALSE     FALSE
#> 4    core_functions   TRUE      TRUE
#> 5             delta  FALSE     FALSE
#> 6          ducklake  FALSE     FALSE
#> 7         encodings  FALSE     FALSE
#> 8             excel  FALSE     FALSE
#> 9               fts  FALSE     FALSE
#> 10               h3   TRUE      TRUE
#> 11           httpfs  FALSE     FALSE
#> 12          iceberg  FALSE     FALSE
#> 13              icu  FALSE     FALSE
#> 14             inet  FALSE     FALSE
#> 15         jemalloc  FALSE     FALSE
#> 16             json  FALSE     FALSE
#> 17       motherduck  FALSE     FALSE
#> 18    mysql_scanner  FALSE     FALSE
#> 19          parquet   TRUE      TRUE
#> 20 postgres_scanner  FALSE     FALSE
#> 21          spatial   TRUE      TRUE
#> 22   sqlite_scanner  FALSE     FALSE
#> 23            tpcds  FALSE     FALSE
#> 24             tpch  FALSE     FALSE
#> 25               ui  FALSE     FALSE
#> 26              vss  FALSE     FALSE
#>                                                                                 install_path
#> 1                                                                                           
#> 2                                                                                           
#> 3                                                                                           
#> 4                                                                                 (BUILT-IN)
#> 5                                                                                           
#> 6                                                                                           
#> 7                                                                                           
#> 8                                                                                           
#> 9                                                                                           
#> 10      /home/runner/.local/share/R/duckdb/extensions/v1.3.2/linux_amd64/h3.duckdb_extension
#> 11                                                                                          
#> 12                                                                                          
#> 13                                                                                          
#> 14                                                                                          
#> 15                                                                                          
#> 16                                                                                          
#> 17                                                                                          
#> 18                                                                                          
#> 19                                                                                (BUILT-IN)
#> 20                                                                                          
#> 21 /home/runner/.local/share/R/duckdb/extensions/v1.3.2/linux_amd64/spatial.duckdb_extension
#> 22                                                                                          
#> 23                                                                                          
#> 24                                                                                          
#> 25                                                                                          
#> 26                                                                                          
#>                                                                           description
#> 1                                          Adds support for autocomplete in the shell
#> 2                                        Provides features that depend on the AWS SDK
#> 3                      Adds a filesystem abstraction for Azure blob storage to DuckDB
#> 4                                                               Core function library
#> 5                                                         Adds support for Delta Lake
#> 6                                Adds support for DuckLake, SQL as a Lakehouse Format
#> 7                                                      All unicode encodings to UTF-8
#> 8                                          Adds support for Excel-like format strings
#> 9                                           Adds support for Full-Text Search Indexes
#> 10              H3 hierarchical hexagonal indexing system for geospatial data, v4.3.0
#> 11               Adds support for reading and writing files over a HTTP(S) connection
#> 12                                                    Adds support for Apache Iceberg
#> 13                   Adds support for time zones and collations using the ICU library
#> 14                               Adds support for IP-related data types and functions
#> 15                                          Overwrites system allocator with JEMalloc
#> 16                                                   Adds support for JSON operations
#> 17                                     Enables motherduck integration with the system
#> 18                                    Adds support for connecting to a MySQL database
#> 19                                 Adds support for reading and writing parquet files
#> 20                                 Adds support for connecting to a Postgres database
#> 21 Geospatial extension that adds support for working with spatial data and functions
#> 22                         Adds support for reading and writing SQLite database files
#> 23                                      Adds TPC-DS data generation and query support
#> 24                                       Adds TPC-H data generation and query support
#> 25                                                           Adds local UI for DuckDB
#> 26                       Adds indexing support to accelerate Vector Similarity Search
#>            aliases extension_version      install_mode installed_from
#> 1                                        NOT_INSTALLED               
#> 2                                        NOT_INSTALLED               
#> 3                                        NOT_INSTALLED               
#> 4                                    STATICALLY_LINKED               
#> 5                                        NOT_INSTALLED               
#> 6                                        NOT_INSTALLED               
#> 7                                        NOT_INSTALLED               
#> 8                                        NOT_INSTALLED               
#> 9                                        NOT_INSTALLED               
#> 10                           efb8a57        REPOSITORY      community
#> 11 http, https, s3                       NOT_INSTALLED               
#> 12                                       NOT_INSTALLED               
#> 13                                       NOT_INSTALLED               
#> 14                                       NOT_INSTALLED               
#> 15                                       NOT_INSTALLED               
#> 16                                       NOT_INSTALLED               
#> 17              md                       NOT_INSTALLED               
#> 18           mysql                       NOT_INSTALLED               
#> 19                                   STATICALLY_LINKED               
#> 20        postgres                       NOT_INSTALLED               
#> 21                           662718c        REPOSITORY           core
#> 22 sqlite, sqlite3                       NOT_INSTALLED               
#> 23                                       NOT_INSTALLED               
#> 24                                       NOT_INSTALLED               
#> 25                                       NOT_INSTALLED               
#> 26                                       NOT_INSTALLED